requirements.txt
openai>=1.14.2
python-dotenv
python-telegram-bot==20.3
requests
httpx
